feat(#81): Tag Mapper 분리
