+++ app/src/main/java/com/ssafy/neegongnaegong/data/mapper/learningrecord/LearningRecordMapper.kt
@@ -1,6 +1,6 @@
 package com.ssafy.neegongnaegong.data.mapper.learningrecord
 
-import com.ssafy.neegongnaegong.data.mapper.TagMapper.toDomain
+import com.ssafy.neegongnaegong.data.mapper.tag.LearningRecordTagMapper.toDomain
 import com.ssafy.neegongnaegong.data.model.learningrecord.request.CreateLearningRecordRequest
 import com.ssafy.neegongnaegong.data.model.learningrecord.request.UpdateLearningRecordRequest
 import com.ssafy.neegongnaegong.data.model.learningrecord.response.GetLearningRecordResponse
